# openapi.yaml
openapi: 3.0.3
info:
  title: DekoHome by Meyran - REST API
  version: 1.0.0
  description: >
    DekoHome by Meyran; modern mobilya ve ev dekorasyonu ürünlerini web ve mobil platformlarda
    kullanıcılarla buluşturan bir e-ticaret uygulamasıdır. Kullanıcılar ürün kataloğunu inceleyebilir,
    kategorilere göre filtreleme yapabilir, ürünleri sepetlerine ekleyebilir, sipariş oluşturabilir ve
    satın aldıkları ürünler için yorum bırakabilir.

    Bu OpenAPI dokümanı; kimlik doğrulama (JWT), kullanıcı profili, kategori yönetimi,
    ürün yönetimi, sepet işlemleri, sipariş süreçleri ve ürün yorumlarını kapsar.
    API; güvenli erişim için JWT tabanlı kimlik doğrulama ile korunur ve yetki gerektiren
    işlemler (ör. admin güncellemeleri) bu kapsamda değerlendirilir.
  contact:
    name: Meyran Proje Ekibi
servers:
  - url: http://localhost:3000/v1
    description: Yerel geliştirme sunucusu
  - url: https://dekohome-api.onrender.com/v1
    description: Domain Linki
...
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Kimlik Doğrulama]
      summary: Üye Olma
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        "201":
          description: Kullanıcı oluşturuldu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          description: Geçersiz istek verisi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Kimlik Doğrulama]
      summary: Giriş Yapma
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        "200":
          description: Giriş başarılı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "401":
          description: E-posta veya şifre hatalı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # KULLANICI PROFİLİ
...
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'JWT tabanlı kimlik doğrulama. "Authorization: Bearer <token>" başlığı eklenmelidir.'
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: "Geçersiz istek"
      required: [message]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      required: [token]
    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          example: "user@example.com"
        password:
          type: string
          example: "StrongPass123!"
        ad:
          type: string
          example: "Ebrar"
        soyad:
          type: string
          example: "Karakoç"
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          example: "user@example.com"
        password:
          type: string
          example: "StrongPass123!"
    User:
      type: object
      properties:
        id:
          type: integer
          example: 101
        email:
          type: string
          example: "user@example.com"
        ad:
          type: string
          example: "Ebrar"
        soyad:
          type: string
          example: "Karakoç"
...